Move biip.gs1{ => _application_identifiers}.GS1ApplicationIdentifier
