+++ docs/api/gs1_application_identifiers.md
@@ -1,0 +1,3 @@
+# `biip.gs1_application_identifiers` - GS1 application identifiers
+
+::: biip.gs1_application_identifiers
