+++ src/biip/gs1_application_identifiers/_application_identifiers.json
