+++ tests/gs1/test_element_strings.py
@@ -6,7 +6,8 @@ import pytest
 
 from biip import ParseError
 from biip.gln import Gln
-from biip.gs1 import GS1ApplicationIdentifier, GS1ElementString
+from biip.gs1 import GS1ElementString
+from biip.gs1_application_identifiers import GS1ApplicationIdentifier
 from biip.gs1_prefixes import GS1CompanyPrefix, GS1Prefix
 from biip.gtin import Gtin, GtinFormat
 from biip.sscc import Sscc
