+++ tests/test_parse.py
@@ -4,7 +4,8 @@ from decimal import Decimal
 import pytest
 
 from biip import ParseResult, parse
-from biip.gs1 import GS1ApplicationIdentifier, GS1ElementString, GS1Message
+from biip.gs1 import GS1ElementString, GS1Message
+from biip.gs1_application_identifiers import GS1ApplicationIdentifier
 from biip.gs1_prefixes import GS1CompanyPrefix, GS1Prefix
 from biip.gtin import Gtin, GtinFormat, Rcn, RcnRegion, RcnUsage
 from biip.sscc import Sscc
